+++ examples/data-sources/datapower_rest_mgmt_interface/data-source.tf
@@ -1,0 +1,3 @@
+
+data "datapower_rest_mgmt_interface" "test" {
+}