+++ examples/resources/datapower_rest_mgmt_interface/resource.tf
@@ -1,0 +1,5 @@
+
+resource "datapower_rest_mgmt_interface" "test" {
+  local_port    = 5554
+  local_address = "0.0.0.0"
+}