+++ internal/provider/datasources/data_source_datapower_rest_mgmt_interface.go
@@ -1,0 +1,140 @@
+// Copyright © 2025 Scott Wiederhold <s.e.wiederhold@gmail.com>
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+// This file is generated "gen/generator.go"
+// !!CHANGES TO THIS FILE WILL BE OVERWRITTEN!!
+
+package datasources
+
+import (
+	"context"
+	"fmt"
+	"regexp"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+	"github.com/scottw514/terraform-provider-datapower/internal/provider/actions"
+	"github.com/scottw514/terraform-provider-datapower/internal/provider/models"
+	"github.com/scottw514/terraform-provider-datapower/internal/provider/tfutils"
+)
+
+var (
+	_ datasource.DataSource              = &RestMgmtInterfaceDataSource{}
+	_ datasource.DataSourceWithConfigure = &RestMgmtInterfaceDataSource{}
+)
+
+func NewRestMgmtInterfaceDataSource() datasource.DataSource {
+	return &RestMgmtInterfaceDataSource{}
+}
+
+type RestMgmtInterfaceDataSource struct {
+	pData *tfutils.ProviderData
+}
+
+func (d *RestMgmtInterfaceDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_rest_mgmt_interface"
+}
+
+func (d *RestMgmtInterfaceDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: "Configure the REST management interface. <p>If you do not assign a TLS profile, the service uses a profile with a self-signed certificate.</p>",
+		Attributes: map[string]schema.Attribute{
+			"provider_target": schema.StringAttribute{
+				MarkdownDescription: tfutils.NewAttributeDescription("Target host to retrieve this data from. If not set, provider will use the top level settings.", "", "").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(1, 128),
+					stringvalidator.RegexMatches(regexp.MustCompile("^[a-zA-Z0-9_-]+$"), "Must match :"+"^[a-zA-Z0-9_-]+$"),
+				},
+			},
+			"enabled": schema.BoolAttribute{
+				MarkdownDescription: "<p>The administrative state of the configuration.</p><ul><li>To make active, set to enabled.</li><li>To make inactive, set to disabled.</li></ul>",
+				Computed:            true,
+			},
+			"user_summary": schema.StringAttribute{
+				MarkdownDescription: "Comments",
+				Computed:            true,
+			},
+			"local_port": schema.Int64Attribute{
+				MarkdownDescription: "Specify the TCP port that the interface monitors. The default value is 5554.",
+				Computed:            true,
+			},
+			"ssl_config_type": schema.StringAttribute{
+				MarkdownDescription: "Custom TLS server type",
+				Computed:            true,
+			},
+			"ssl_server": schema.StringAttribute{
+				MarkdownDescription: "Custom TLS server profile",
+				Computed:            true,
+			},
+			"ssl_sni_server": schema.StringAttribute{
+				MarkdownDescription: "Custom TLS SNI server profile",
+				Computed:            true,
+			},
+			"local_address": schema.StringAttribute{
+				MarkdownDescription: "<p>Enter a host alias or the IP address that the service listens on. Host aliases can ease migration tasks among appliances.</p><ul><li>0 or 0.0.0.0 indicates all configured IPv4 addresses.</li><li>:: indicates all configured IPv4 and IPv6 addresses.</li></ul><p><b>Attention:</b> For management services, the value of 0.0.0.0 or :: is a security risk. Use an explicit IP address to isolate management traffic from application data traffic.</p>",
+				Computed:            true,
+			},
+			"dependency_actions": actions.ActionsSchema,
+		},
+	}
+}
+
+func (d *RestMgmtInterfaceDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	d.pData = req.ProviderData.(*tfutils.ProviderData)
+}
+
+func (d *RestMgmtInterfaceDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	d.pData.Mu.Lock()
+	defer d.pData.Mu.Unlock()
+
+	var data models.RestMgmtInterface
+	resp.Diagnostics.Append(req.Config.Get(ctx, &data)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if data.ProviderTarget.ValueString() != "" && !d.pData.Client.ValidTarget(data.ProviderTarget.ValueString()) {
+		resp.Diagnostics.AddError("Invalid provider_target", fmt.Sprintf(`Target %q is not defined in the provider's 'targets' block. Available targets: %v`, data.ProviderTarget.ValueString(), d.pData.Client.GetTargetNames()))
+		return
+	}
+
+	path := data.GetPath()
+
+	res, err := d.pData.Client.Get(path, data.ProviderTarget)
+	resFound := true
+	if err != nil {
+		if !strings.Contains(err.Error(), "status 404") {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))
+			return
+		} else {
+			resFound = false
+		}
+	}
+	if resFound {
+		data.FromBody(ctx, `RestMgmtInterface`, res)
+	}
+
+	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
+}