+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## UNRELEASED
+
+- Added `RestMgmtInterface` endpoint
+
 ## v0.15.0
 
 - Bump github.com/hashicorp/terraform-plugin-docs from 0.23.0 to 0.24.0 in the terraform-plugin group (PR20)